Frack Blood Particle Created!
The base blood particle for the Frack is now here! I'll probably modify this one so that it fits for each of the units. (I.E small blood cloud for Sappers, big for Boss Units).

It isn't applied to any units at the moment but can be seen in the effects demo scene in the VFX folder.